extract hex formatting extension method GetFormattedCsHexArray to the GeneratePatchTargetHashes project
